# bank-reactive-microservices/accounts-service/openapi-accounts.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
  /api/v1/accounts/{id}:
    get:
      tags:
      - account-controller
      operationId: findById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountResponse"
    put:
      tags:
      - account-controller
      operationId: update
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountResponse"
    delete:
      tags:
      - account-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /api/v1/movements:
    get:
      tags:
      - movement-controller
      operationId: findAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovementResponse"
    post:
      tags:
      - movement-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovementRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MovementResponse"
  /api/v1/accounts:
    get:
      tags:
      - account-controller
      operationId: findAll_1
      parameters:
      - name: customerId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
    post:
      tags:
      - account-controller
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AccountResponse"
  /reports/{clientId}:
    get:
      tags:
      - report-controller
      operationId: report
      parameters:
      - name: clientId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: startDate
        in: query
        required: true
        schema:
          type: string
          format: date-time
      - name: endDate
        in: query
        required: true
        schema:
          type: string
          format: date-time
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MovementReportRow"
  /api/v1/movements/{id}:
    get:
      tags:
      - movement-controller
      operationId: findById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MovementResponse"
    delete:
      tags:
      - movement-controller
      operationId: delete_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
components:
  schemas:
    UpdateAccountRequest:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        type:
          type: string
          enum:
          - SAVINGS
          - CHECKING
        balance:
          type: number
        status:
          type: boolean
      required:
      - balance
      - customerId
      - status
      - type
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        accountNumber:
          type: string
        type:
          type: string
          enum:
          - SAVINGS
          - CHECKING
        balance:
          type: number
        status:
          type: boolean
        createdAt:
          type: string
          format: date-time
    CreateMovementRequest:
      type: object
      properties:
        accountId:
          type: integer
          format: int64
        type:
          type: string
        amount:
          type: number
      required:
      - accountId
      - amount
      - type
    MovementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        movementDate:
          type: string
          format: date-time
        type:
          type: string
        amount:
          type: number
        balance:
          type: number
        status:
          type: boolean
    CreateAccountRequest:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        accountNumber:
          type: string
        type:
          type: string
          enum:
          - SAVINGS
          - CHECKING
        balance:
          type: number
        status:
          type: boolean
      required:
      - accountNumber
      - balance
      - customerId
      - status
      - type
    MovementReportRow:
      type: object
      properties:
        date:
          type: string
          format: date-time
        customerId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        accountNumber:
          type: string
        accountType:
          type: string
        openingBalance:
          type: number
        status:
          type: boolean
        movementAmount:
          type: number
        movementType:
          type: string
        availableBalance:
          type: number